reorganized widget files
