+++ Routine-Machine/lib/Views/pages/HabitDetailPage.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../../constants/Palette.dart' as Palette;
-import '../../BottomNavBar.dart';
+import '../components/BottomNavBar.dart';
 
 class HabitDetailPage extends StatelessWidget {
   @override
